Fix the double `!again` problem
Previously, when two players entered the command `!again` in a short
span of time, a new game was set (hands where distributed to players)
then immediatly erased and reset.

This was done by redifining how phases of the game work.
Moreover, it is now impossible tho reset the game with `!again` during the play
phase.